+++ packages/components/releaseNotes/components.md
@@ -1,6 +1,13 @@
 # @labkey/components
 Components, models, actions, and utility functions for LabKey applications and pages.
 
+### version TBD
+*Released*: TBD
+* @labkey/components package bundle optimizations
+    - webpack.config.js: add back in new IgnorePlugin(/^\.\/locale$/, /moment$/)
+    - webpack.config.js: remove "enzyme" realted objects from bundle via "externals"
+    - split initUnitTestMocks() into a separate testHelperMocks.tsx file so it isn't referenced via ./src/index.ts
+
 ### version 2.2.0
 *Released*: 19 February 2021
 * Item 8335: Add LabKey product navigation icon and menu to NavigationBar.tsx
