+++ packages/components/src/internal/components/QueryGridPanel.spec.tsx
@@ -20,7 +20,8 @@ import { List } from 'immutable';
 import { TESTS_ONLY_RESET_DOM_COUNT } from '../util/utils';
 
 import { getStateQueryGridModel, QueryGridModel, SchemaQuery, QueryGridPanel } from '../..';
-import { initUnitTestMocks, registerDefaultURLMappers, sleep } from '../testHelpers';
+import { registerDefaultURLMappers, sleep } from '../testHelpers';
+import { initUnitTestMocks } from '../testHelperMocks';
 
 beforeAll(() => {
     initUnitTestMocks();
