+++ packages/components/src/internal/components/entities/SingleParentEntityPanel.spec.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 
 import { mount } from 'enzyme';
 
-import { initUnitTestMocks } from '../../testHelpers';
+import { initUnitTestMocks } from '../../testHelperMocks';
 
 import { QueryGridPanel, SelectInput } from '../../..';
